MainTabs [nfc]: Combine with `MainScreenWithTabs`.
It's always been confusing for these to be separate things. With
React Navigation 5's component-based API, it seems quite natural to
combine them.
